requirements.txt
eventlet~=0.33.3
Flask~=2.2.3
flask-cors~=6.0.1
Flask-SocketIO~=5.3.6
SQLAlchemy~=2.0.19
opencv-python~=4.8.1.78
numpy~=1.26.4
ultralytics~=8.0.221
requests~=2.31.0
pandas~=2.1.4
plotly~=5.17.0
schedule~=1.2.0
python-dotenv~=1.0.0
